Added "php functions.php", Fixed genQueryProperties()
Added php functions.php to the library now, it contains some little
function snippets

By using themyEmpty() function from functions.php it allows for values
to be equal to "0" and not be considered empty.  This was causing some
issues while trying to grab a row with values that equalled zero.