New functionality
ORM:
Changed out $uniqueUnusedFirstProperty to $tableStructureDelimiter
because it's shorter and more memorable, overall it has better
readability and practicality to it.

Added $sqlConstants, beacause I was running into an issue where
CURRENT_TIMESTAMP couldn't be wrapped in single quotes, which is also
why I ended up adding the new functions to  "php functions.php".  I'm
not sure what else shouldn't be surrounded in quotes but when/if more
arrises all that needs to take place is the addition to the array for
the new content.

I would though like to properly code this so that any value in a query
would first call a method that would determine wether or not the value
shouldn't be wrapped in single quotes.
This would cut down on the redundant ugly
if ( in_array($value, $this->sqlConstants) )
code that I used, it was kind of a late night messy fix that works, that
I did not even need to do, but I felt that since I ran into this issue,
others shouldn't need to do a hacky fix, which is the whole point of
this library.

On line 338, I did add a comment stating this; that I do not think that
the key value to a table structure will ever be a constant, because it
defeates the purpose of a key being unique, as well as constants are
constant, even though CURRENT_TIMESTAMP's value is not constant this
situation coudn't really exist.
I will surely clean this up when I get around to refractoring this how I
would like it to be.

php functions:
Added two one/two new functions to php functions.php
They generate the current proper timestamp or datetime for mysql