New Methods, small changes
getCollection, getCollectionDataOnly, customClause, public $db

You can now have it return a collection of objects w/o the class
overhead, so that you can easily dump it out into json if you wanted to.

Changed it so that you can use a custom clause with "useProperties=true"
The clause will use the properties to generate a clause and then append
the custom clause to the end of it.  This works great for appending
"ORDER BY"

I made the $db property public, I believe that was to simplify some
coding, but I think it ended up being more of a hassle than it did good,
so it might need to be removed if it is obsolete/useless